+++ src/GIMI-ModManager.WinUI/Models/Settings/CharacterGallerySettings.cs
@@ -13,4 +13,10 @@ public class CharacterGallerySettings
     public bool IsSingleSelection { get; set; } = true;
 
     public bool IsNavPaneOpen { get; set; } = false;
+
+    public bool CanDeleteDialogPrompt { get; set; } = true;
+
+    public bool SortByDescending { get; set; } = false;
+
+    public string? SortingMethod { get; set; }
 }