+++ src/GIMI-ModManager.WinUI/Validators/PreConfigured/ModsFolderValidator.cs
@@ -1,5 +1,4 @@
 ﻿using FluentValidation;
-using GIMI_ModManager.WinUI.Services;
 using PathPicker = GIMI_ModManager.WinUI.ViewModels.SubVms.PathPicker;
 
 namespace GIMI_ModManager.WinUI.Validators.PreConfigured;